# backend/src/main/resources/api/openapi.yaml
openapi: 3.0.3
info:
  title: Bau Platform API
  description: API for connecting construction companies with temporary workers
  version: 1.0.0
  contact:
    name: Bau Platform Team
    email: support@bau-platform.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.bau-platform.com/api/v1
    description: Production server
  - url: https://staging-api.bau-platform.com/api/v1
    description: Staging server
  - url: http://localhost:8080/api/v1
    description: Local development server

# Global security - can be overridden at operation level
security:
  - BearerAuth: []

# Import paths from separate files

# Import components from separate files
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from AWS Cognito
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
          description: Error timestamp
          example: "2024-01-15T10:30:00Z"
        status:
          type: integer
          description: HTTP status code
          example: 400
        error:
          type: string
          description: Error type
          example: "Bad Request"
        message:
          type: string
          description: Error message
          example: "Validation failed"
        path:
          type: string
          description: Request path
          example: "/api/v1/bedarf"
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
                description: Field name
                example: "datumVon"
              message:
                type: string
                description: Field-specific error message
                example: "Start date must be in the future"
...
    UserResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: User ID
          example: "123e4567-e89b-12d3-a456-426614174000"
        username:
          type: string
          description: Username
          example: "john.doe@bau-ag-grabs.ch"
        email:
          type: string
          format: email
          description: Email address
          example: "john.doe@bau-ag-grabs.ch"
        firstName:
          type: string
          description: First name
          example: "John"
        lastName:
          type: string
          description: Last name
          example: "Doe"
        betriebId:
          type: string
          format: uuid
          description: Company ID
          example: "123e4567-e89b-12d3-a456-426614174001"
        betriebName:
          type: string
          description: Company name
          example: "Bau AG Grabs"
        role:
          type: string
          enum: [ADMIN, BETRIEB]
          description: User role
          example: "BETRIEB"
        status:
          type: string
          enum: [AKTIV, INACTIV]
          description: User status
          example: "AKTIV"
        emailVerified:
          type: boolean
          description: Email verification status
          example: true
        createdAt:
          type: string
          format: date-time
          description: Account creation timestamp
          example: "2024-01-15T10:30:00Z"
        updatedAt:
          type: string
          format: date-time
          description: Last update timestamp
          example: "2024-01-15T10:30:00Z"
    CreateUserRequest:
      type: object
      required:
        - username
        - email
        - firstName
        - lastName
        - betriebId
        - role
      properties:
        username:
          type: string
          description: Username
          example: "john.doe@bau-ag-grabs.ch"
        email:
          type: string
          format: email
          description: Email address
          example: "john.doe@bau-ag-grabs.ch"
        firstName:
          type: string
          description: First name
          example: "John"
        lastName:
          type: string
          description: Last name
          example: "Doe"
        betriebId:
          type: string
          format: uuid
          description: Company ID
          example: "123e4567-e89b-12d3-a456-426614174001"
        role:
          type: string
          enum: [ADMIN, BETRIEB]
          description: User role
          example: "BETRIEB"
    UpdateUserRequest:
      type: object
      properties:
        username:
          type: string
          description: Username
          example: "john.doe@bau-ag-grabs.ch"
        email:
          type: string
          format: email
          description: Email address
          example: "john.doe@bau-ag-grabs.ch"
        firstName:
          type: string
          description: First name
          example: "John"
        lastName:
          type: string
          description: Last name
          example: "Doe"
        betriebId:
          type: string
          format: uuid
          description: Company ID
          example: "123e4567-e89b-12d3-a456-426614174001"
        role:
          type: string
          enum: [ADMIN, BETRIEB]
          description: User role
          example: "BETRIEB"
        status:
          type: string
          enum: [AKTIV, INACTIV]
          description: User status
          example: "AKTIV"
    UpdateUserProfileRequest:
      type: object
      properties:
        firstName:
          type: string
          description: First name
          example: "John"
        lastName:
          type: string
          description: Last name
          example: "Doe"
        email:
          type: string
          format: email
          description: Email address
          example: "john.doe@bau-ag-grabs.ch"
    ChangePasswordRequest:
      type: object
      required:
        - currentPassword
        - newPassword
      properties:
        currentPassword:
          type: string
          description: Current password
          example: "oldPassword123"
        newPassword:
          type: string
          description: New password
          example: "newPassword123"
    UserListResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/UserResponse'
        totalElements:
          type: integer
          description: Total number of elements
          example: 100
        totalPages:
          type: integer
          description: Total number of pages
          example: 5
        currentPage:
          type: integer
          description: Current page number
          example: 1
        pageSize:
          type: integer
          description: Page size
          example: 20
        hasNext:
          type: boolean
          description: Has next page
          example: true
        hasPrevious:
          type: boolean
          description: Has previous page
          example: false 
paths:
...
  /users:
    get:
      operationId: getUsers
      summary: List users
      description: Get paginated list of users with optional filtering
      tags:
        - Users
      parameters:
        - name: page
          in: query
          description: Page number (1-based)
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: size
          in: query
          description: Page size
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: betriebId
          in: query
          description: Filter by company ID
          required: false
          schema:
            type: string
            format: uuid
        - name: role
          in: query
          description: Filter by role
          required: false
          schema:
            type: string
            enum: [ADMIN, BETRIEB]
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserListResponse'
        '401':
          description: Unauthorized
  
    post:
      operationId: createUser
      summary: Create new user
      description: Create a new user
      tags:
        - Users
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - insufficient permissions
  /users/{id}:
    get:
      operationId: getUserById
      summary: Get user by ID
      description: Retrieve a specific user by their ID
      tags:
        - Users
      parameters:
        - name: id
          in: path
          required: true
          description: User ID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: User details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '404':
          description: User not found
        '401':
          description: Unauthorized
  
    put:
      operationId: updateUser
      summary: Update user
      description: Update an existing user
      tags:
        - Users
      parameters:
        - name: id
          in: path
          required: true
          description: User ID
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: User updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Invalid request data
        '401':
          description: Unauthorized
        '404':
          description: User not found
  
    delete:
      operationId: deleteUser
      summary: Delete user
      description: Delete a user by their ID
      tags:
        - Users
      parameters:
        - name: id
          in: path
          required: true
          description: User ID
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: User deleted successfully
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - insufficient permissions
        '404':
          description: User not found
  /users/profile:
    get:
      operationId: getCurrentUserProfile
      summary: Get current user profile
      description: Get the profile for the currently authenticated user
      tags:
        - Users
      responses:
        '200':
          description: User profile retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          description: Unauthorized
  
    put:
      operationId: updateCurrentUserProfile
      summary: Update current user profile
      description: Update the profile for the currently authenticated user
      tags:
        - Users
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserProfileRequest'
      responses:
        '200':
          description: User profile updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Invalid request data
        '401':
          description: Unauthorized
  /users/change-password:
    post:
      operationId: changePassword
      summary: Change password
      description: Change password for the currently authenticated user
      tags:
        - Users
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
      responses:
        '200':
          description: Password changed successfully
        '400':
          description: Invalid request data
        '401':
          description: Unauthorized 